requirements.txt
aiohttp==3.6.3
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.11.8
chardet==3.0.4
discord.py==1.5.1
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
lxml==4.6.1
mccabe==0.6.1
multidict==4.7.6
Pillow==8.0.1
pkg-resources==0.0.0
pylint==2.6.0
requests==2.25.0
shakeabuse==1.0.2
six==1.15.0
SQLAlchemy==1.3.20
toml==0.10.2
urllib3==1.26.1
wrapt==1.12.1
yarl==1.5.1
